requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg==3.1.16
djangorestframework-simplejwt==5.3.1
